Fix a bug in album art detection
New album art files were not being assigned to albums and copied into
the appropriate location.

Normally, the scanner determines what album a certain art file is for
by looking for files in the same directory as the art file and pulling
the correct album_id from the database.

The database query was using `LIKE ‘home/icecast/…%’` and was not
finding any matching album_ids because the query was missing the
leading slash in the directory name.

This patch will send the directory name with the leading slash in the
query, allowing the database to return the appropriate album_ids for
album art.