Prevent scanner from incorrectly disabling songs
In cases where there are two music directories and the name of one is a
prefix of the other, the scanner would incorrectly disable some songs.

For example, if the scanner picked up a modification event on the
directory `/music/Metroid`, it would disable all songs in the directory
`/music/MetroidPrime` because it searches the database for any files
with a prefix that matches the original directory.

This patch introduces a trailing path separator to the database search
so such songs are not disabled.